requirements.txt
fastapi==0.110.1
fastapi-mail==1.4.1
motor==3.4.0
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
fastapi-mqtt==2.1.1
python-socketio==5.11.2
uvicorn==0.29.0
starlette~=0.37.2
APScheduler~=3.10.4
cryptography~=42.0.5
aiofiles~=23.2.1